Modularise postgres queries
